Refactor file structure & renaming functions
